feat(book): add error handling for invalid ids

+++ src/main/java/com/springpractice/bookstore/controller/BookController.java
@@ -1,5 +1,6 @@
 package com.springpractice.bookstore.controller;
 
+import com.springpractice.bookstore.exceptions.ResourceNotFoundException;
 import com.springpractice.bookstore.model.Book;
 import com.springpractice.bookstore.repository.BookRepository;
 import org.springframework.http.HttpStatus;
@@ -28,10 +29,9 @@ public class BookController {
 
     @PutMapping("/{id}")
     public Book updateBook(@PathVariable long id, @RequestBody Book book) {
-        Book existingBook = bookRepository.findById(id).orElse(null);
-        if (existingBook == null) {
-            return null;
-        }
+        Book existingBook = bookRepository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Book with id " + id + " not found")
+        );
         book.setId(id);
         return bookRepository.save(book);
     }
@@ -39,11 +39,16 @@ public class BookController {
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteBook(@PathVariable long id) {
+        if (!bookRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Book with id " + id + " not found");
+        }
         bookRepository.deleteById(id);
     }
 
     @GetMapping("/isbn/{isbn}")
     public Book getBooksByIsbn(@PathVariable String isbn) {
-        return bookRepository.findByIsbn(isbn).orElse(null);
+        return bookRepository.findByIsbn(isbn).orElseThrow(
+                () -> new ResourceNotFoundException("Book with ISBN (" + isbn + ") not found")
+        );
     }
 }
